+++ src/PhpSpreadsheet/Worksheet/Drawing.php
@@ -89,10 +89,11 @@ class Drawing extends BaseDrawing
      * @param string $path File path
      * @param bool $verifyFile Verify file
      * @param ?ZipArchive $zip Zip archive instance
+     * @param null|callable(string):bool $isWhitelisted
      *
      * @return $this
      */
-    public function setPath(string $path, bool $verifyFile = true, ?ZipArchive $zip = null, bool $allowExternal = true): static
+    public function setPath(string $path, bool $verifyFile = true, ?ZipArchive $zip = null, bool $allowExternal = true, ?callable $isWhitelisted = null): static
     {
         $this->isUrl = false;
         if (preg_match('~^data:image/[a-z]+;base64,~', $path) === 1) {
@@ -102,12 +103,24 @@ class Drawing extends BaseDrawing
         }
 
         $this->path = '';
+        if ($zip instanceof ZipArchive) {
+            $zipPath = explode('#', $path)[1];
+            $locate = @$zip->locateName($zipPath);
+            if ($locate !== false) {
+                if ($this->isImage($path)) {
+                    $this->path = $path;
+                    $this->setSizesAndType($path);
+                }
+            }
         // Check if a URL has been passed. https://stackoverflow.com/a/2058596/1252979
-        if (filter_var($path, FILTER_VALIDATE_URL) || (preg_match('/^([\w\s\x00-\x1f]+):/u', $path) && !preg_match('/^([\w]+):/u', $path))) {
+        } elseif (filter_var($path, FILTER_VALIDATE_URL) || (preg_match('/^([\w\s\x00-\x1f]+):/u', $path) && !preg_match('/^([\w]+):/u', $path))) {
             if (!preg_match('/^(http|https|file|ftp|s3):/', $path)) {
                 throw new PhpSpreadsheetException('Invalid protocol for linked drawing');
             }
             if (!$allowExternal) {
+                return $this;
+            }
+            if ($isWhitelisted !== null && !$isWhitelisted($path)) {
                 return $this;
             }
             // Implicit that it is a URL, rather store info than running check above on value in other places.
@@ -142,15 +155,6 @@ class Drawing extends BaseDrawing
                         }
                         unlink($filePath);
                     }
-                }
-            }
-        } elseif ($zip instanceof ZipArchive) {
-            $zipPath = explode('#', $path)[1];
-            $locate = @$zip->locateName($zipPath);
-            if ($locate !== false) {
-                if ($this->isImage($path)) {
-                    $this->path = $path;
-                    $this->setSizesAndType($path);
                 }
             }
         } else {