+++ src/PhpSpreadsheet/Writer/Pdf/Tcpdf.php
@@ -8,6 +8,10 @@ use PhpOffice\PhpSpreadsheet\Writer\Pdf;
 
 class Tcpdf extends Pdf
 {
+    protected bool $writeHeader = false;
+
+    protected bool $writeFooter = false;
+
     /**
      * Create a new PDF Writer instance.
      *
@@ -66,8 +70,8 @@ class Tcpdf extends Pdf
         $pdf->SetMargins($printMargins->getLeft() * 72, $printMargins->getTop() * 72, $printMargins->getRight() * 72);
         $pdf->SetAutoPageBreak(true, $printMargins->getBottom() * 72);
 
-        $pdf->setPrintHeader(false);
-        $pdf->setPrintFooter(false);
+        $pdf->setPrintHeader($this->writeHeader);
+        $pdf->setPrintFooter($this->writeFooter);
 
         $pdf->AddPage();